tools/requirements-semantic.txt
openai>=1.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0

